+++ src/main/resources/application.properties
@@ -63,3 +63,6 @@ naver.client-secret=${NAVER_CLIENT_SECRET}
 #Google login Secret
 google.client-id=${GOOGLE_CLIENT_ID}
 google.client-secret=${GOOGLE_CLIENT_SECRET}
+
+#Constant
+image-prefix="https://storage.googleapis.com/plan-storage/";
